requirements.txt
pbr>=0.6,<1.0
# Core Requirements
Django>=1.4,<1.7
django_compressor>=1.3
MySQL-python>=1.2.5
Pillow>=2.3
httplib2>=0.8
qiniu>=6.1.3
markdown>=2.4
django-filter>=0.7
DjangoRestless
pygments
Fabric==1.6.0
django-tinymce==1.5.2
qrcode==4.0.4
six==1.5.2
djangorestframework==2.3.6
django-htmlmin>=0.6.2
django-grappelli>=2.5.1
django-colorful>=0.1.3
python-memcached>=1.53
sinaweibopy>=1.1.3
django_compressor>=1.3
django-appconf>=0.6
versiontools>=1.9.1
BeautifulSoup>=3.2.1
redis>=2.8.0
hiredis>=0.1.1
pymongo>=2.6.3
django-debug-toolbar>=1.0
lxml==3.3.0beta4 #mac 可能需要升级xcode：xcode-select --install ,  服务器也可能需要升级sudo yum install libxslt-devel libxml2-devel  libxslt-devel
django_pickling>=0.1
django-cacheops>=1.3.1